usb: dwc3: dwc3-msm-core: Modify PORTSC check for device connection status
While in host mode, the dwc3_set_phy_speed_flags() API will check the USB
root hub's PORTSC value to determine the connection type, i.e. SS/HS/FS/LS
device connected.  During configuration of the wakeup interrupts, this flag
is utilized to set the proper wakeup triggers.

In order to address a corner case when a device fails to enumerate and is
left connected to the RH, the condition was changed to check for the
PORT_CONNECT bit versus PORT_ENABLE.  In this situation, the PORTSC value
shows as:
    dwc3_set_phy_speed_flags reg = 0x6e1

The PORT_PE value is cleared as part of the suspend sequence, so the speed
flag is set to not reflect the speed of the device connected.  This would
result in the DP/DM interrupts being programmed to look for a level HIGH
condition, which will not occur if the device is disconnected.

The change to utilize PORT_CONNECT led to a side effect where if the user
did a quick plug out/in of the device, the speed flag PORTSC check would
improperly set the wakeup flag to look for a level LOW condition.  This
would not occur leading to the device to fail to enumerate:
    dwc3_set_phy_speed_flags reg = 0xa0206e1

To fix both conditions, add an additional check for the PORT_CSC status, as
if there is a device connected while executing the speed flags API, then
that means a connect seqeuence has to be serviced, and should configure the
wakeup interrupts to look for level HIGH.  In the other case, the PORT_CSC
has been already handled and cleared by the hub driver, and if still
connected, should be configuring interrupts to look for a level LOW.

Change-Id: Icfad3e4b3611346d13ac805bbb5f96084f02ff7b
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>